requirements.txt
LAMP-core
datetime
numpy
pandas
altair
tzwhere
geopy
similaritymeasures
fastdtw
sklearn
pytz